feat: create Layout component and refactor CustomerLoginPage to use LoginForm with improved structure
